requirement.txt
streamlit
pillow
tweepy
python-dotenv
